Refactor: ViewHolder 계층 구조 선언
